sparc: make ka compile

+++ assemblers/ka/a.h
@@ -1,7 +1,7 @@
 #include <u.h>
 #include <libc.h>
 #include <bio.h>
-#include "../kc/k.out.h"
+#include <obj/k.out.h>
 
 #ifndef	EXTERN
 #define	EXTERN	extern
@@ -127,7 +127,7 @@ void	newfile(char*, int);
 Sym*	slookup(char*);
 Sym*	lookup(void);
 void	syminit(Sym*);
-long	yylex(void);
+int	yylex(void);
 int	getc(void);
 int	getnsc(void);
 void	unget(int);
